fix: refactor code to use optimized protobuf marshal
We should always use wrapper around `proto.Marshal` and
`proto.Unmarshal` to make sure vtprotobuf optimization is picked up
correctly.
